flask-app/pyproject.toml
[tool.poetry]
name = "cbl-web-tool"
version = "0.1.0"
description = ""
authors = ["Alex Swindler <alex.swinder@nrel.gov>", "Nicholas Long <nicholas.long@nrel.gov>", "Roger Braun <roger.braun@nrel.gov>", "Rahman Mian <rahman.mian@nrel.gov>"]
readme = "README.md"
# should probably rename utils to cbl_workflow, then place
# the utils in that folder.
packages = [{include = "utils"}]

[tool.poetry.dependencies]
python = ">=3.9, <3.13"

flask = "^3.0.3"
flask-cors = "^4.0.1"
geopandas = "^0.14.3"
mercantile = "^1.2.1"
openpyxl = "^3.1.5"
pnnl-buildingid = {git = "ssh://git@github.com/seed-platform/buildingid.git", rev = "master"}
requests = "^2.31.0"
shapely = "^2.0.3"
street-address = "0.4.0"
usaddress = "0.5.10"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"

[build-system]
# Need to provide the build system information for the package to be built
requires = ["poetry-core", "setuptools", "wheel"]
build-backend = "poetry.core.masonry.api"
